moves leader and development cards to feature lib
