+++ libs/lorenzo/feature-browse/src/lib/cards/development/development.component.scss
