+++ libs/lorenzo/feature-browse/src/lib/cards/leader/leader.component.scss
