+++ libs/lorenzo/ui/src/index.ts
@@ -1,4 +1,4 @@
+export * from './lib/card/card.component';
+export * from './lib/card/card-header/card-header.component';
 export * from './lib/cards/card-template.directive';
 export * from './lib/cards/cards.component';
-export * from './lib/development/development.component';
-export * from './lib/leader/leader.component';
